规范化异常处理

+++ src/main/java/top/hiasenna/community/controller/PublishController.java
@@ -28,7 +28,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,Model model) {
+    public String edit(@PathVariable(name = "id") Long id,Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
         model.addAttribute("description",question.getDescription());
@@ -46,7 +46,7 @@ public class PublishController {
     public String doPublish(
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
-            @RequestParam(value = "id",required = false)Integer id,
+            @RequestParam(value = "id",required = false)Long id,
             @RequestParam(value = "tag",required = false) String tag, HttpServletRequest request, Model model) {
         model.addAttribute("title", title);
         model.addAttribute("description", description);
