+++ src/main/java/top/hiasenna/community/controller/QuestionController.java
@@ -6,7 +6,6 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import top.hiasenna.community.dto.QuestionDTO;
-import top.hiasenna.community.mapper.QuestionMapper;
 import top.hiasenna.community.service.QuestionService;
 
 /**
@@ -22,7 +21,7 @@ public class QuestionController {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name="id")Integer id, Model model){
+    public String question(@PathVariable(name="id")Long id, Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         //累加阅读数
         questionService.incView(id);
